fix: support calling hook function multiple times
Previously, if the hook function was called more than once, only the
callback given in the last call would be active. So if you either called
the same hook function multiple times or if another module in your
dependency tree had its own copy of the `require-in-the-middle` module,
the last one to call `hook` would win.

With this change, the hooks will be called in series, where the last one
to call the `hook` function will be the first to be called when a module
is required (LIFO).

Fixes #11