feat!: use `require.cache` for caching patched exports for non-core modules (#63)
This allows a user to delete from `require.cache` to trigger a re-load
of a module (and a re-run of the hook's `onrequire`) on next `require()`.

This also removes the `.cache` field from the Hook object. It was never
documented, but it wasn't prefixed with `_` to hint at being private. As
well, some tests used it, which might have given the impression it could
be used externally. I'm calling this a semver-major change to give a
heads up to users that might have been using it.

Fixes: #61
Obsoletes: #57