fix: Don't crash if there are require.cache shenanigans in play
There can, in general, be something else in the process that is
manipulating `require.cache` -- e.g. `@babel/register` -- such that when
a module that RITM is hooking, it might *not* be in `require.cache`
after `require()`ing it.

Before this change, this would result in RITM crashing on an assert.
With this change RITM will fall back to caching the hooked module's exports
internally -- as it used to do in versions before v7 -- instead of
on the `require.cache[filename]` entry.

Fixes: https://github.com/elastic/require-in-the-middle/issues/72